Move CLI interface into cleanlab_studio.cli module
This patch prepares the code for the addition of a Python library
interface. With that change, only stable APIs should be exported by the
`cleanlab_studio` package, with the exception of the
`cleanlab_studio.cli` and `cleanlab_studio.internal` modules. For now,
basically all the code has been moved into the `cli` module; as the
Python library interface is developed, common functionality will be
pulled out of there and into `internal`.
