+++ cleanlab_studio/__init__.py
@@ -1,13 +1,1 @@
-from .version import (
-    __version__ as __version__,
-    SCHEMA_VERSION as SCHEMA_VERSION,
-    MIN_SCHEMA_VERSION as MIN_SCHEMA_VERSION,
-    MAX_SCHEMA_VERSION as MAX_SCHEMA_VERSION,
-    SETTINGS_VERSION as SETTINGS_VERSION,
-    MIN_SETTINGS_VERSION as MIN_SETTINGS_VERSION,
-    MAX_SETTINGS_VERSION as MAX_SETTINGS_VERSION,
-)
-from . import login
-from . import cleanset
-from . import dataset
-from . import util
+from .version import __version__ as __version__
