+++ cleanlab_studio/cli/__init__.py
@@ -1,0 +1,1 @@
+from cleanlab_studio.cli.main import cli as main
