+++ cleanlab_studio/cli/classes/json_dataset.py
@@ -4,7 +4,7 @@ import pandas as pd
 import ijson
 
 from .dataset import Dataset
-from cleanlab_studio.types import RecordType
+from cleanlab_studio.cli.types import RecordType
 
 
 class JsonDataset(Dataset):
