+++ cleanlab_studio/cli/dataset/commands.py
@@ -1,6 +1,6 @@
 import click
-from cleanlab_studio.dataset.schema import schema
-from cleanlab_studio.dataset.upload import upload
+from cleanlab_studio.cli.dataset.schema import schema
+from cleanlab_studio.cli.dataset.upload import upload
 
 
 @click.group(help="upload datasets, generate & validate dataset schema")
