+++ cleanlab_studio/cli/util.py
@@ -7,9 +7,9 @@ import pathlib
 import pandas as pd
 from typing import Optional, Dict, Any, List, Generator
 
-from cleanlab_studio.classes import CsvDataset, JsonDataset, ExcelDataset
-from cleanlab_studio.classes.dataset import Dataset
-from cleanlab_studio.types import (
+from cleanlab_studio.cli.classes import CsvDataset, JsonDataset, ExcelDataset
+from cleanlab_studio.cli.classes.dataset import Dataset
+from cleanlab_studio.cli.types import (
     RecordType,
     DatasetFileExtension,
     ImageFileExtension,
