+++ tests/row_processing/test_datetime_feature_type.py
@@ -1,11 +1,11 @@
 from typing import Tuple, Optional
 
-from cleanlab_studio.dataset import (
+from cleanlab_studio.cli.dataset import (
     RowWarningsType,
     FeatureType,
     DataType,
 )
-from cleanlab_studio.types import RecordType
+from cleanlab_studio.cli.types import RecordType
 from tests.row_processing.utils import process_record_with_fields
